% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/20055R.R
\name{alg}
\alias{alg}
\title{to calculate the min edit distance of 2 string}
\usage{
alg(word1, word2)
}
\arguments{
\item{word1}{one string}

\item{word2}{the other string}
}
\description{
to calculate the min edit distance of 2 string
}
\examples{
\dontrun{
f=function(x) c(x[1]^3-x[2]-1,x[1]^3-x[2]^2-1)
alg(c(1,2),f)
}
}
