% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/20055R.R
\name{s}
\alias{s}
\title{An function to judge whether a netural number is a prime number}
\usage{
s(n)
}
\arguments{
\item{n}{the need number}
}
\value{
a string to see whether a number is aprime number and its factor
}
\description{
to judge whether a netural number is a prime number
}
\examples{
\dontrun{
f=function(x) c(x[1]^3-x[2]-1,x[1]^3-x[2]^2-1)
f(c(1,2))
s(c(1,2))
}
}
